+++ libraries/strings/sibling_deps.ts
@@ -1,0 +1,1 @@
+export { pickItem } from '../arrays/src/pickItem.ts';